% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooserInput.R
\name{chooserInput}
\alias{chooserInput}
\title{Starts the shiny app}
\usage{
chooserInput(inputId, leftChoices, rightChoices, size = 5, multiple = FALSE)
}
\description{
Call this function to start the shiny app.
}
\keyword{internal}
