% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template_code.R
\name{get_template_code}
\alias{get_template_code}
\title{Get Templates}
\usage{
get_template_code(type, markup = "LaTeX")
}
\arguments{
\item{type}{Options: schoice, num, mchoice, string, cloze}

\item{markup}{Can be Latex or Rmarkdown}
}
\value{
The templates of schoice, num, mchoice, string or cloze exercises
}
\description{
Get templates to show in the editor viewer
}
\examples{
get_template_code('mchoice', 'LaTeX')
}
