% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooserInput.R
\name{chooserInput}
\alias{chooserInput}
\title{Chooser Input}
\usage{
chooserInput(inputId, leftChoices, rightChoices, size = 5, multiple = FALSE)
}
\arguments{
\item{multiple}{}
}
\value{

}
\description{
Chooser Input
}
